chore: move logger formatter to its own file
